Little restructure of the OutputBlock and CodeFile class
