+++ E-Paper-Image-Creator/E-Paper-Image-Creator.csproj
@@ -24,7 +24,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <None Update="eau.png">
+    <None Update="Examples\eau.png">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
     <None Update="img.png">
