+++ E-Paper-Image-Creator/ParseException.cs
@@ -1,0 +1,25 @@
+﻿using System;
+using System.Runtime.Serialization;
+
+namespace Zkwip.EPIC
+{
+    [Serializable]
+    internal class ParseException : Exception
+    {
+        public ParseException()
+        {
+        }
+
+        public ParseException(string? message) : base(message)
+        {
+        }
+
+        public ParseException(string? message, Exception? innerException) : base(message, innerException)
+        {
+        }
+
+        protected ParseException(SerializationInfo info, StreamingContext context) : base(info, context)
+        {
+        }
+    }
+}
